requirements.txt
django==1.10.5
social-auth-core==1.1.0
social-auth-app-django==1.0.1
django-tenant-schemas==1.6.11